Clearly instruct to delete and insert new history for feature replication
When the Data Integrator creates calculates new features it drops all
old values and re-creates the entire feature history. This is done so as
not to keep old values that have been filtered away due to a new FtQl
definition, or incoming corrections on the data events from the source.
We will have to require our customers to do the same if  they whish to
have a consistent history of features in their own systems. This commit
makes this mechanism very explicit.

This change might require attention from customers who have already
implemented a feature sync mechanism without paying attention to this
detail. At the time this commit was created no customer has made a full
implementation yet. If that is no longer the case upon merging of this
commit, further communication with customers is needed.